Fix checking whether condition cf is applied to current object
This fix a bug preventing correct processing of conditioned CF in
display mode. What was wrongly checked was whether the condition CF was
applied to the current object (eg. a RT::Ticket) while the condition CF
is actually applied the parent class of current object (eg. a
RT::Queue). Tests were wrongly OK because they apply on a ticket and
queue that have the same id.